sites-holders: fix a race condition which deadlocked the task
When the task start happens a bit late, the queue might already contain
the Empty_Holder added by Load to trigger the expiration time
computation. In that case, the empty update holder made the task wait,
while the queue considered the task as active.

Now the holder emptiness is no longer overloaded, an extra boolean
argument instead communicates whether to wait or work on the update.